refactor(api): extract zip logic into its own file
